requirements.txt
certifi==2019.9.11
cffi==1.13.2
cryptography==2.8
cycler==0.10.0
future==0.18.2
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
python-telegram-bot==12.2.0
pytz==2019.3
six==1.13.0
tornado==6.0.3
